fix: PHPStan level 3 — zero errors
- Fix return types: navigation badges cast to string, CheckoutController
  returns Checkout, ProductWaitlistController union return type
- Fix StripeWebhookController return types to match Cashier parent
- Fix Holiday::getDaysAwayAttribute() cast diffInDays to int
- Fix CapacityLimit::forDate() broken query and orphan cast entries
- Fix RecipeCostCalculator foreach on json cast (add @property)
- Fix MenuController nested Collection generic (->all())
- Add @property for Recipe::$ingredients json cast
- Bump PHPStan to level 3

+++ app/Models/Holiday.php
@@ -73,7 +73,7 @@ class Holiday extends Model
 
     protected function getDaysAwayAttribute(): int
     {
-        return now()->diffInDays($this->date, false);
+        return (int) now()->diffInDays($this->date, false);
     }
 
     protected function getStartPrepByAttribute(): Carbon
